refactor input form
