+++ frontend/components/lv4/WorkDetail.tsx
@@ -1,6 +1,7 @@
-import { FC, FormEvent } from "react";
+import { FC } from "react";
 import { Work } from "model";
 import { WorkImage } from "components/lv1/WorkImage";
+import { WorkOriginalForm } from 'components/lv3/WorkOriginalForm'
 import styled from "styled-components";
 import Constants from "styles/Constants";
 import { useWorkOriginals } from "hooks/work/useWorkOriginals";
@@ -14,22 +15,6 @@ const WorkTitle = styled.div`
   font-weight: ${Constants.FONT_WEIGHT.BOLD};
 `
 
-const WorkOriginalInput: FC<{ work: Work }> = ({ work }) => {
-  const onSubmit = (event: FormEvent<HTMLFormElement>) => {
-    event.preventDefault()
-    console.log('on submit')
-  }
-
-  return (
-    <form onSubmit={ onSubmit }>
-      <input type="text" placeholder="原作種別"/>
-      <input type="number" placeholder="何巻？"/>
-      <input type="text" placeholder="アニメ何話時点？"/>
-      <button type="submit">送信</button>
-    </form>
-  )
-}
-
 const WorkOriginal: FC<{ work: Work }> = ({ work }) => {
   const { loading, originals } = useWorkOriginals(work)
   if (loading) return (
@@ -39,7 +24,7 @@ const WorkOriginal: FC<{ work: Work }> = ({ work }) => {
   if (originals.length === 0) return (
     <div>
       no data
-      <WorkOriginalInput work={ work } />
+      <WorkOriginalForm work={ work } />
     </div>
   )
 
@@ -49,7 +34,7 @@ const WorkOriginal: FC<{ work: Work }> = ({ work }) => {
         <div key={`original_${i}`}>{JSON.stringify(original)}</div>
       ))}
 
-      <WorkOriginalInput work={ work } />
+      <WorkOriginalForm work={ work } />
     </div>
   )
 }
